Update pyproject.toml
The new setuptools has deprecated the ways in which we added entry points, scripts, and dependencies in the past.
This had a curious effect in which `pip install .` would not install dependencies or console scripts.
We could switch to putting this all into the pyproject.toml, but for now I used the approved workaround
of listing these as "dynamic" in the project section of the pyproject.toml.