Change query to use catalog table only
After discussion with Derek, it looks like we should *not* try to query
in the standard way, since this may miss events which were low significance online
but high signifance offline. Rather this change will assume that the list of superevents
which should be queried is exactly that which is in the gwtc table. It must therefore
be confirmed with search folks that this table won't include e.g. every low significance event.