Refactor `fetch_gracedb_information()`
This creates a new function with the same name as the original, reproducing the behavior.
The reason for this non-functional change is to make the code *much* more readable.
This should substantially improve maintainability and make future upgrades much easier,
especially for the upcoming catalog updates.

This commit still includes the old function under the name
`old_fetch_gracedb_information()` for comparison purposes, but once checks have been
done to reasonable satisfaction this can be removed.

Finally this new version is compatible with proper unit-testing, though that
would be a significant additional contribution.