ENH: Speed-up catalog operation, and capture CWB events.
Catalog operations were slow because I was calling gdb.gwtc_get() each iteration of the core loop,
and this operation is apparently very slow (25 s as opposed to 0.1 s for an event call).
Refactoring allows this to only be called once, hence making it essentially as fast as the online operation.

Slight changes were also made to the handling of CWB events, so that all the CWB events in the pipeline can be
captured correctly. These didn't throw any errors during handling, so the parameters used (search value either
'AllSky' or 'BBH') are apparently fine, but we should double check this with the CWB developers / allsky group.