ENH: Always store GraceDB preferred event, even in catalog operation
Asimov needs a stable definition of preferred event, especially for runs beginning
with online data. Accordingly, we should always fetch these preferred events.
This commit adds logic so that we do so.

Tracking whether this event is in the catalog table will be done by a new schema
element included in a separate commit, and this will be updated by code introduced
with the schema v3 monitor feature.